SEC-440/project2_webapp/requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.3
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
protobuf==4.21.12
Werkzeug==2.3.7
zipp==3.17.0